Ignore null release body
Using the SOLIDSoftworks/semver-tags action will result in a release
created that doesn't contain anything for the release body. The
consequences is that the updated release body will appear to contain the
string `null` followed by a newline and then the release body as
specified for this action.

In such a case it appears to make more sense to ignore the existing
release body should it be currently `null` instead of implicitly
converting it to a string for the updated release body.